+++ Sources/StreamVideoSwiftUI/CallView/VideoParticipantsView.swift
@@ -35,8 +35,8 @@ public struct VideoParticipantsView<Factory: ViewFactory>: View {
                 ParticipantsFullScreenLayout(
                     viewFactory: viewFactory,
                     participant: first,
+                    call: viewModel.call,
                     size: availableSize,
-                    pinnedParticipant: $viewModel.pinnedParticipant,
                     onViewRendering: onViewRendering,
                     onChangeTrackVisibility: onChangeTrackVisibility
                 )
@@ -44,17 +44,17 @@ public struct VideoParticipantsView<Factory: ViewFactory>: View {
                 ParticipantsSpotlightLayout(
                     viewFactory: viewFactory,
                     participant: first,
+                    call: viewModel.call,
                     participants: Array(viewModel.participants.dropFirst()),
                     size: availableSize,
-                    pinnedParticipant: $viewModel.pinnedParticipant,
                     onViewRendering: onViewRendering,
                     onChangeTrackVisibility: onChangeTrackVisibility
                 )
             } else {
                 ParticipantsGridLayout(
                     viewFactory: viewFactory,
+                    call: viewModel.call,
                     participants: viewModel.participants,
-                    pinnedParticipant: $viewModel.pinnedParticipant,
                     availableSize: availableSize,
                     orientation: orientation,
                     onViewRendering: onViewRendering,
@@ -73,23 +73,20 @@ public struct VideoCallParticipantModifier: ViewModifier {
     @State var popoverShown = false
     
     var participant: CallParticipant
-    @Binding var pinnedParticipant: CallParticipant?
-    var participantCount: Int
+    var call: Call?
     var availableSize: CGSize
     var ratio: CGFloat
     var showAllInfo: Bool
     
     public init(
         participant: CallParticipant,
-        pinnedParticipant: Binding<CallParticipant?>,
-        participantCount: Int,
+        call: Call?,
         availableSize: CGSize,
         ratio: CGFloat,
         showAllInfo: Bool
     ) {
         self.participant = participant
-        _pinnedParticipant = pinnedParticipant
-        self.participantCount = participantCount
+        self.call = call
         self.availableSize = availableSize
         self.ratio = ratio
         self.showAllInfo = showAllInfo
@@ -104,7 +101,7 @@ public struct VideoCallParticipantModifier: ViewModifier {
                         HStack {
                             ParticipantInfoView(
                                 participant: participant,
-                                isPinned: participant.id == pinnedParticipant?.id
+                                isPinned: participant.isPinned
                             )
                             
                             if showAllInfo {
@@ -124,17 +121,50 @@ public struct VideoCallParticipantModifier: ViewModifier {
                     }
                     
                     if popoverShown {
-                        Button {
-                            if participant.id == pinnedParticipant?.id {
-                                self.pinnedParticipant = nil
-                            } else {
-                                self.pinnedParticipant = participant
+                        VStack(spacing: 16) {
+                            if !participant.isPinnedRemotely {
+                                PopoverButton(
+                                    title: pinTitle,
+                                    popoverShown: $popoverShown
+                                ) {
+                                    if participant.isPinned {
+                                        Task {
+                                            try await call?.unpin(
+                                                sessionId: participant.sessionId
+                                            )
+                                        }
+                                    } else {
+                                        Task {
+                                            try await call?.pin(
+                                                sessionId: participant.sessionId
+                                            )
+                                        }
+                                    }
+                                }
                             }
-                            popoverShown = false
-                        } label: {
-                            Text(participant.id == pinnedParticipant?.id ? L10n.Call.Current.unpinUser : L10n.Call.Current.pinUser)
-                                .padding(.horizontal)
-                                .foregroundColor(.primary)
+                            
+                            if call?.state.ownCapabilities.contains(.pinForEveryone) == true {
+                                PopoverButton(
+                                    title: pinForEveryoneTitle,
+                                    popoverShown: $popoverShown
+                                ) {
+                                    if participant.isPinnedRemotely {
+                                        Task {
+                                            try await call?.unpinForEveryone(
+                                                userId: participant.userId,
+                                                sessionId: participant.id
+                                            )
+                                        }
+                                    } else {
+                                        Task {
+                                            try await call?.pinForEveryone(
+                                                userId: participant.userId,
+                                                sessionId: participant.id
+                                            )
+                                        }
+                                    }
+                                }
+                            }
                         }
                         .padding()
                         .modifier(ShadowViewModifier())
@@ -150,6 +180,24 @@ public struct VideoCallParticipantModifier: ViewModifier {
                 }
             }
     }
+    
+    @MainActor
+    private var participantCount: Int {
+        call?.state.participants.count ?? 0
+    }
+    
+    private var pinTitle: String {
+        participant.isPinned
+            ? L10n.Call.Current.unpinUser
+            : L10n.Call.Current.pinUser
+    }
+    
+    private var pinForEveryoneTitle: String {
+        participant.isPinnedRemotely
+            ? L10n.Call.Current.unpinForEveryone
+            : L10n.Call.Current.pinForEveryone
+    }
+    
 }
 
 public struct VideoCallParticipantView: View {
@@ -265,4 +313,22 @@ public struct SoundIndicator: View {
             .streamAccessibility(value: participant.hasAudio ? "1" : "0")
     }
     
+}
+
+struct PopoverButton: View {
+    
+    var title: String
+    @Binding var popoverShown: Bool
+    var action: () -> ()
+    
+    var body: some View {
+        Button {
+            action()
+            popoverShown = false
+        } label: {
+            Text(title)
+                .padding(.horizontal)
+                .foregroundColor(.primary)
+        }
+    }
 }