+++ StreamVideoSwiftUITests/CallView/ParticipantsGridLayout_Tests.swift
@@ -2,12 +2,36 @@
 // Copyright © 2023 Stream.io Inc. All rights reserved.
 //
 
+@testable import StreamVideo
 @testable import StreamVideoSwiftUI
 import SnapshotTesting
 import XCTest
 
 @MainActor
 final class ParticipantsGridLayout_Tests: StreamVideoUITestCase {
+    
+    nonisolated private lazy var callController = CallController_Mock(
+        defaultAPI: DefaultAPI(
+            basePath: "test.com",
+            transport: httpClient as! HTTPClient_Mock,
+            middlewares: []
+        ),
+        user: StreamVideo.mockUser,
+        callId: callId,
+        callType: callType,
+        apiKey: "123",
+        videoConfig: VideoConfig(),
+        cachedLocation: nil
+    )
+    
+    override func setUpWithError() throws {
+        try super.setUpWithError()
+        
+        let streamVideo = StreamVideo.mock(httpClient: httpClient, callController: callController)
+        streamVideoUI = StreamVideoUI(streamVideo: streamVideo)
+    }
+    
+    private lazy var call = streamVideoUI?.streamVideo.call(callType: callType, callId: callId)
     
     private func gridSize(for participantsCount: Int) -> CGSize {
         let heightDivider = CGFloat((participantsCount == 2 || participantsCount == 3) ? participantsCount : 1)
@@ -18,8 +42,8 @@ final class ParticipantsGridLayout_Tests: StreamVideoUITestCase {
         for count in gridParticipants {
             let layout = ParticipantsGridLayout(
                 viewFactory: TestViewFactory(participantLayout: .grid, participantsCount: count),
+                call: call,
                 participants: ParticipantFactory.get(count, withAudio: true),
-                pinnedParticipant: .constant(nil),
                 availableSize: gridSize(for: count),
                 orientation: .portrait,
                 onViewRendering: {_,_ in },
@@ -33,8 +57,8 @@ final class ParticipantsGridLayout_Tests: StreamVideoUITestCase {
         for count in gridParticipants {
             let layout = ParticipantsGridLayout(
                 viewFactory: TestViewFactory(participantLayout: .grid, participantsCount: count),
+                call: call,
                 participants: ParticipantFactory.get(count, withAudio: false),
-                pinnedParticipant: .constant(nil),
                 availableSize: gridSize(for: count),
                 orientation: .portrait,
                 onViewRendering: {_,_ in },
@@ -49,8 +73,8 @@ final class ParticipantsGridLayout_Tests: StreamVideoUITestCase {
             let count = gridParticipants.last!
             let layout = ParticipantsGridLayout(
                 viewFactory: TestViewFactory(participantLayout: .grid, participantsCount: count),
+                call: call,
                 participants: ParticipantFactory.get(count, connectionQuality: quality),
-                pinnedParticipant: .constant(nil),
                 availableSize: gridSize(for: count),
                 orientation: .portrait,
                 onViewRendering: {_,_ in },
@@ -62,10 +86,16 @@ final class ParticipantsGridLayout_Tests: StreamVideoUITestCase {
     
     func test_grid_participantsSpeaking_snapshot() {
         for count in gridParticipants {
+            let participants = ParticipantFactory.get(count, speaking: true)
+            var dict = [String: CallParticipant]()
+            for participant in participants {
+                dict[participant.id] = participant
+            }
+            callController.update(participants: dict)
             let layout = ParticipantsGridLayout(
                 viewFactory: TestViewFactory(participantLayout: .grid, participantsCount: count),
-                participants: ParticipantFactory.get(count, speaking: true),
-                pinnedParticipant: .constant(nil),
+                call: call,
+                participants: participants,
                 availableSize: gridSize(for: count),
                 orientation: .portrait,
                 onViewRendering: {_,_ in },